django_projects/SaasApp/backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-crispy-forms==1.14.0
django-resized==1.0.2
django-tinymce==3.5.0
html2text==2020.1.16
idna==3.4
numpy==1.24.1
Pillow==9.4.0
psycopg2==2.9.5
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
uuid==1.30
